EOS-17934: added benchmark name/parameters to the task list. Fixed path to Perfline log file
