+++ performance/PerfLine/roles/perfline_setup/files/webui/config.py
@@ -1,7 +1,6 @@
 perfline_script_path = '/root/perfline/wrapper/perfline.py'
 artifacts_dir = '/var/perfline'
-logfile = '/var/perfline/log/nohup.log'
+logfile = '/var/log/perfline.log'
 
-#server_ep = {'host': '10.230.247.173', 'port': 5005}
 server_ep = {'host': '0.0.0.0', 'port': 8005}
